fix(vite): only rewrite fully static src values into asset imports
The static-asset rewrite scans compiled render output for src keys and
hoists alias-matched string values into imports. A bound :src compiles
to a concatenation — src: '/images/colors-switch-' + (color ? 'on' :
'off') + '.jpg' — and the scanner took the FIRST string literal as the
whole value, importing half a filename (or a bare directory), which
rolldown fails to load (#3945). @vitejs/plugin-vue only rewrites static
attribute values.

parse_src_candidate now accepts a candidate only when the property ends
right after the literal's closing quote (comma, closing brace/paren, or
EOF follows) — a literal inside a larger expression is a runtime value
and stays untouched. A bare src key must also start at an identifier
boundary, so data_src-style props no longer match.

Closes #3945
